Refactor uninstall_all_but_latest to allow passing in ui
